Added 'sprint mode', 'ultra mode'
Tweaks to rank system. It was too easy to get an 'M' ranking for speed
or combos in ultra mode. The 'perfect player' always eliminated 1 line
at a time. With practice, it's also quite feasible to keep a combo going
indefinitely, or at least for a very very long time.