Fixed edge case for pieces sliding into nooks.
'Instant nook DAS' now activates when the player is pushing a piece into
a wall, regardless of whether or not the piece has technically attempted
to move into a wall yet.

With the old logic pieces could slide past nooks in some annoying cases.
For example, sliding an O-piece into the right three columns would work
OK because the O-piece spawns on the right side of the playfield. But
sliding an O-piece into the left three columns would fail, because it
would successfully move and not activate 'instant nook DAS'.