Customer grows up to 10x bigger as you build a big combo.
Camera also adjusts as the customer grows. I increased the size of the
RestaurantScene to handle the case where the camera is way zoomed out.

The customer currently shrinks when a combo ends, but this isn't permanent.
I made it so the customer doesn't shrink when a level ends; this feels better
for 'ultra' mode