Ending a combo pans to a new customer.
The restaurant now contains three customers. When the camera pans away,
the customer who was just fed shrinks to their original size.
Eventually, they'll instead be replaced with a new customer.

Added a CustomerViewDemo for showing off the "pan to new customer"
logic, as well as tracking down bugs related to feeding the wrong bug
and things like that.

This involved some very messy code and kludges for having shadow and
outline sprites track each other across different scenes. This is
inevitable until we resolve issues #30 and #32 to reduce the amount of
work in maintaining hierarchies of interconnected sprites which are all
drawing the same simple object.

After rearranging the customers and furniture into reusable pieces, I
noticed a bug where food was being drawn behind the table, instead of
being hurled over top of it. I rearranged the objects so that food is
drawn correctly, but the resulting design is very ugly.

Fixed an NPE when OutlineSprite's parent was undefined.