GLES3. Fixed 'on_game_ended' console warning. Disabled camera scaling.
The camera scaling for big customers looked cool, but the second and
third customer didn't scale properly, and the camera drifting off to
the right. I couldn't figure out an easy fix.

GLES2 was producing some warnings, so I upgraded to GLES3. This makes us
unable to export to HTML5. We can change this back, it's not a big deal.