Customers now appear in 16 different colors.
The current approach uses a complex 'RgbPalette.shader' shader, which
relies on an unusual style of sprite. The unusual sprite involves three
red, green and blue channels which each include data for custom colors,
and the remaining 'black' channel is used for lines. The shader converts
these red/green/blue colors to arbitrary colors, e.g aquamarine, puce
and taupe.

If, in the future, more than 3 colors are needed, this shader can be
applied to multiple sprites layered on top of one another. Hopefully
that's sufficient, because even as kludgy as this current approach is,
it involved a lot of trial and error, and about 2-3 failed technical
approaches. The benefits of the current approach are that the heavy
lifting can be offloaded to graphics cards. The drawbacks are that the
sprite files need to be ugly images with bright colors.

One failed approach was a naive shader-based approach which recolored the existing sprites based on whether they matched certain arbitrary colors (e.g change forest green to orange.) This failed because the shader was applied after sprites were scaled, so the user could clearly see the aliased pixels in the resulting algorithm.

Another failed approach involved loading the texture images and
replacing the pixels one-by-one. This looked good but was exceptionally
slow, taking about 15 seconds to recolor a customer even with the
current limited number of animations. The mouth-sheet.png files is
8215x1026, for example, and computers are not vet fpushry good at doing
something 8,000,000 times.