Customers make sounds when they enter, eat, and leave.
When they enter, a door chimes and sometimes they say 'hello'. If the
player gets a big combo, the customer will says 'mmm!' or 'oooh!' in between
bites. When the player breaks their combo, the customer says 'check please!"

The customer's sounds are AudioStreamPlayer2Ds with a spatial element.
The door is on the left side of the store, so when the door chimes or
customers say 'hello', the sounds come from the left. When customers say
'mmm!' or 'check please!' then the sound comes from close to the center.

I also tried to make the other customer sounds lean a little to the left, since
the customers are on the left side of the screen. I figure kitchen sounds will
be panned right and eating sounds will be panned left.

There's some logic to make sure customers don't say 'hello' 'goodbye'
too frequently.