Fixed puzzle freezing when prelock was interrupted.
The prelock state can result in two states; writing the piece to the
playfield or going back to the move state if the player smushes or
performs a lock reset. If these two state changes overlapped, bugs
resulted such as infinite loops or freezes.

I've refactored the surrounding code to minimize the number of places
where we manage the piece state using 'set_state'. The states themselves
now monitor 'apply_player_input' and change the state themselves.