Enlarged 3D objects to reduce physics glitches.
There was some unusual code in turbo.gd to reduce jitter. Controlling a box which is 2 units high causes 0.2 units of jitter in Godot, which is very noticable as it is 10% of the object's height. I attempted to reduce this jitter by snapping the character's Y coordinate to the nearest 0.2, but this was not a perfect fix. Some bouncing was still visible occasionally. It also didn't work if Turbo ever landed on a surface that wasn't a perfect multiple of 0.2.

I've made all 3D objects 900% bigger to reduce these physics glitches. Controlling a box which is 20 units high makes the jitter less visible. I also removed the snapping behavior for Turbo's movement.

Fixed lighting artifacts; orthogonal lights were introducing weird stripes on textures in editor