Turbo slips from ledges in 3D mode.
This prevents her from hanging off a ledge in mid-air, which was jarring
for two reasons: she was sort of suspended in mid-air megaman style with
maybe a toe touching the ledge or maybe not, which looked strange. But also, there was a visible
gap between her shadow and the ledge's shadow.

I also refactored a few references to use Vector2.ZERO and Vector3.ZERO
for readability.