Fixed 2D graphical regressions; borders, camera
The 'wall metal' texture had a visible seam at the bottom. Changing its
size to 256 (a power of 2) appears to make the seam disappear. Perhaps all
of our textures should be powers of 2. Many already are.

Fixed a visible outline around the tentacle mouth when customers ate.

Fixed a broken animation which was introduced by renaming the upper-case
animation name "Fat" to the lower-case "fat".