Puzzle timing changes; shorter pause when making boxes.
Instead of pausing for 30 frames whether making a box or clearing a line, the
pauses are now vastly different. boxes only result in 50% of the delay of
a line clear. Overall this results in a much more frenetic feel to the
game. I increased the line clear delay to compensate but it might still be too
fast.

Added new 'DELAYS' cheat to see puzzle mode timing data.

Fixed bug where rank calculator could tell the player they scored an
impossible 22.5 points per line, if they finished a level with fewer than
8 line clears.