Changed piece timings; faster but more lenient
The overall goal of these timing changes was:

1. Bring the overall max/min blocks per minute in line with other similar games
2. Allow faster players to play faster on easier difficulties
3. Allow slower players more leniency on easier difficulties
4. Have a smooth difficulty curve (except for the sudden 1G/20G increases)

The frame delay when creating a box or clearing a line has been decreased, especially on easier levels. This allows fast players to play faster.

The number of lock resets and floor kicks has been increased. This makes piece movement more lenient. I've also increased the lock delay from 30 frames to 40 frames for the middle difficulties.

The top speed of the crazy difficulties now remains unchanged for the last few levels. THis allows expert players to find and stick to a rhythm.