Added HTML5 export (2D parts only)
HTML5 does not support threads, so I modified ResourceCache to support a
non-threading loading mechanism for HTML5.

The 3D stuff causes my browser to lock up. I'll have to investigate this
further, but it may just be infeasible for HTML5 to do the 3D parts of the
game. For now I've removed the '3D world' button from the scenario menu.