New grade scale; eliminated D and C ranks.
This new scale follows the progression of our eventual star unlocks.
'B' and 'A' grades will eventually be displayed as dots and triangles. These
will be displayed to the user and help them gauge their progress. Unbeknownst
to them at the start of the game, they have no impact on progress.

S grades will eventually be displayed as stars, like the stars of a
restaurant. Stars will be a prerequisite to hire new employees, grow your
restaurant and increase your income. All of these steps are optional but they
help to progress through the game faster.

These grades