Slower lock times for novices. Faster DAS for experts.
A playtester reported that the lock time of 40 frames felt too oppressive
for what's meant to be a casual game. I've slowed it down to 60 frames for
the beginner levels (0-9)

I found the DAS reduction from FA to FB to be jarring when progressing
through expert difficulties; the pieces felt out of control with only 5 frames
of DAS. I've decreased it to 5 frames for FA as well, so now it feels out of
control all the time. Ha ha.

We can let players adjust their own DAS if they hate it. 5 frames feels OK
to me now that I've practiced it.

Added 'unlock' cheat to scenario select menu.