New u-piece 'shaft kick'
This piece kick was missing, which meant you could drop a u-piece down a
2-cell shaft and be unable to rotate it at the bottom.

I've added this kick and rearranged the existing kicks in an attempt to
keep the u-piece kicks in line with the j and l pieces.

Changed test-piece-kicks framework to distinguish between a piece which
didn't rotate, and a piece which rotated but didn't move.