Save data is backwards-compatible.
Save data from v0.0517 is now transformed to the latest format.

On the surface it's a waste of time to spend 2-3 hours writing code to
maintain save compatibility, when there is literally a single user.
However I think this will continue to be a problem in future versions of
the game, so it's important to establish a backwards compatibility
framework early.

Removed JSON formatting for player save data. This caused a noticable
stutter (~300 ms) when clearing a level.