Expanded/improved intro tutorial. New 'stack stack' section.
Now includes fourth 'snack stack' phase of the tutorial where pleces are
arranged vertically, and the player is told to combine squish moves and
snack boxes to score points.

Player no longer scores points during the beginning phases of the
tutorial. Their score and time is reset during the fourth phase, which is
the only phase that matters.

Player is no longer prompted 'start' and 'back' during tutorial. If they
fail or quit the tutorial, they're prompted 'start'. If they complete the
tutorial, they're prompted 'quit'.

Instead of Playfield having a weird 'read_only' property, it now has a
more versatile 'undo_last_piece' function. This function was necessary for
the new tutorial phase since we can't tell if they've made a block until
the piece is written to the playfield.

Turbo's introductory 'Let's dive right in' dialog now appears immediately,
the player does not need to hit the 'Start' button. This helps reduce the
density of messages during the tutorial.

Dialog is now cleared between phases of the tutorial. Otherwise it was
possible for the tutorial dialog to end up several pages behind if the
player was too quick.

Added default button focus for main screen, splash screen. Fixed skewed
scenario screen, it was offset a few pixels to the side.