Scenario history is saved more efficiently.
The previous design which saved the newest 1,000 (!) scores for each mode had
two flaws.

1. Each score takes about 500 bytes of disk space, so 1,000 scores would
take up 500k of memory and disk space for a single scenario.

2. In the unlikely change that a player played a single scenario more than
1,000 times, the previous strategy would always kick out the oldest record
which could be the player's high score. It would take a great player 5-6 hours
of playing nothing but Ultra Normal to finally kick out their oldest
score, so this is unlikely but it could happen.

The new design saves a maximum of 7 scores: the player's newest score,
their top 3 daily scores, and the top 3 all-time scores.

Extracted ScenarioHistory from PlayerData.