High score system no longer relies on 'ultra-' prefix
The old logic looked at a scenario name to see if it started with
'ultra-', and those scenarios were compared differently. This is a poor
design because it requires all goal-based scenarios to start with the word
'ultra-'.

Now, we take any scenarios with a 'score' goal and give them a different
'compare' property. For now this can be '+score' or '-seconds', but in the
future other scenarios could have other definitions of what it means to
'do a good job at a level'. Some might compare how many or how few customers
you served, reaching a target score in as few lines as possible, or
maximizing the number of boxes made. These could all be represented in
data as something like '+customers', '-customers', '-lines' or
'+box_score'.