Refactored SentenceLabel to use only one label.
I switched it to this design because 'FontFitLabel' seemed clean, but
after implementation of this and investigation of the NametagManager, I'm
not convinced which design is cleaner. It's nice to see the different
labels in the editor, and also gives the flexibility like NametagManager
where you can have labels with different sizes, positions and fonts.

Pulled ChatAppearance constants to top-level to make them easier to
access.