Simplified PuzzleScore signals. 'chat-choices' group.
PuzzleScore had four different signals for different aspects of the score
changing, which was overkill.

Refactored 'chat choice' logic to use button groups. Button groups are
basically intended for this case, of "there's a bunch of of nodes in the
scene tree of this one type, and i want to do something with them'