Fixed bug where creatures had three arms when puzzled
This bug occurred because when 'set_orientation' was called, their arms
would revert to their forward-facing state. However they had an additional
'emote anim' arm sprite which continued to hold their chin.

We no longer revert to the default 'facing in a direction' animation state
if we're simply turning to face left or right. We only revert if we're
going from facing north to facing south or vice-versa.