2D overworld camera zooms in during conversations.
I initially tried to develop this in a way that mirrored the 3D camera, by
having two cameras in different places and toggling which was active.
However unlike 3D, there is no 2D concept of 'smoothly switch from this
target to that target', and the community-suggested solution appears to be
to create a third 'tweening camera' to temporarily lerp between the two
other cameras. Implementing this clumsy solution of two cameras and a
temporary tweening camera was glitchy and involved a lot of custom code,
some of which is available on issue #352.

This single camera approach has a risk of violating SRP and bloating into a
huge class if we have additional events or areas with custom camera behavior.
But, it seems to be the design that's best supported for 2D.