Increased visibility of focused buttons. Themed level editor.
This makes the game easier to play with the controller. Before, it was
sometimes hard to see where your cursor was.

The level editor buttons and components are now themed.

The slider tick is still subtle when it's focused, this will not be easy
to change. We would need to create a custom icon for it.