Changed rank-3d scenario; more level-ups, higher score target
The old 3D scenario had a low score target because it was treated as a
'survival' scenario. I think I initially graded it as survival because it's
fast -- but all of the levels are fast. 'survival' levels are long -- it's
harder to maintain a strong level of play for 10 minutes than 5 minutes.

Having more level-ups leans into its cool gimmick.