Added android export. export_presets.cfg is now generated via a script.
export_presets.cfg include keystore passwords, so it cannot be in version
control. These passwords are now generated via a shell script. This shell
script also updates the version numbers for the project. It should be run
before a release.

Developers who wish to export the project will need to run a shell script
to generate this export_presets.cfg file. They will also need to maintain
a /secrets/ folder which contains keystores and keystore passwords. This
process will be documented in the wiki.