Moved Puzzle hud items into CanvasLayer. Tutorial restart fix.
Puzzle now collects its hud items (chalkboard, messages, buttons) in a
PuzzleHud CanvasLayer. This ensures they're not overwritten by visual
effects like lights, and also makes it easier to add future HUD items (e.g
mobile controls)

If the user got to the last "Earn $100" part of the tutorial and hit
'ESC', they would be taken to the ultra-normal scenario. I've changed it
so that they're taken to the start of the tutorial.