Improved license summary. Added Godot to license.
I've scrutinized each asset and license to ensure they use a
commercial-compatible license.

I'm including a summary of the license, to decide what we will include in
our credits as the game is finished. '$$$' must be included in the game's
credits. '$$' might be included in the game's credits, depending on how
heavily you used it, or whether they're a 'warm fuzzy' project (Gut,
godotengine.org) versus a 'big company' (Google, cheap-audio-drivers.biz)