Added overworld controls.
The overworld controls include some controller-style buttons which appear
similar to the puzzle mode, but with a completely different implementation
backed by a new 'Eightway' node. This node is more customizable and versatile
than the puzzle mode's touch controls. More specifically, it allows for
diagonal presses and misses fewer inputs.

The overworld controls also have new touch controls for conversations.
Rather than requiring the user to use an emulated joystick and buttons,
they can tap the screen or touch the dialog option they want to choose.
This required the implementation of a new TouchTranslator script, which is
quite generic and should see reuse throughout other places in the game.