Puzzle mode touchscreen controls now handle diagonal presses.
Removed old TouchControlSet, TouchControlButton. The idea of using four
TouchscreenButtons for a controller-like input was the easiest thing, but
it does not work very well. The EightWay component works much more
effectively.

Diagonal presses are important for certain techniques, such as flipping
pieces for instant DAS, sliding pieces into nooks, and doing the 'squish
drop' combo move. We might disable this by default, but it needs to be
supported for expert players.