Touch controls can be configured: size, sensitivity, arrangement
Three basic arrangements are available: a simple one for beginners, an
ambidextrous one for people who want to drop with both hands, and a
crazy one for hardcore players who don't mind a learning curve.

Fixed bug when dragging from one eightway to another. Because the
eightways were marking touch events as handled, a drag event from one eightway
to the other resulted in the first eightway not receiving the "touch event is
out of range" input event.