HTML5 issues; wasm filename, slow loading
The HTML5 export was looking for 'v0.wasm' instead of 'v0.0621.wasm'.
Additionally itch.io wants the html file named 'index.html'. I've just
renamed the file to 'turbofat.html' for now, and the extra rename is a
manual step.

The HTML5 export was taking about 5 seconds to load, because it was only
loading one resource every frame. I've changed it to load as many
resources as it can in 0.1 seconds, as a compromise between speed and
responsiveness.