Bug: Piece wouldn't squish sometimes if rotated
Before, the code tracked the 'squish source pos' and 'squish target pos'
and if a piece moved, we recalculated whether a squish move was possible.
This was flawed because a the player might rotate a piece into a position
where a squish move was possible, in which case we would not allow them to
squish the piece, because we had already calculated it was impossible in a
different orientation.

The new code instead tracks a 'squish state' instead of a 'squish source
position'. This state is reset any time the piece moves or rotates.