New version of HipHop04. New song plays twice as often.
If we add a new song every month or two and randomly shuffle them, we'll
hear the old one for 6 months but the newest one for only 1 month. It
makes sense to weight the older songs so they come up less often in the
shuffle.

Eventually, these songs will be worked into different areas of the game, and
this will be moot anyways.