Fixed 'esc' behavior when exiting puzzles.
If you launched a scenario and immediately hit 'esc', the game would quit
to the main menu with the house music still running. You can no longer
exit a scenario unless the 'Quit' button is visible.

If you were in the level editor and hit 'esc', the game would exit the
level, but would also pop up a settings menu. Exiting the scenario now
marks the input as 'handled' to prevent other input handlers from picking
it up.