Fixed visual issues with north-facing creatures
There was a strange floating shape (the 'neck blend') over the player when
facing north on the overworld. This was caused because the 'fat-nw'
animation was never being played, so the neck blend was never moving.

I also made the creature's arm shadows look a little less silly when
facing north. Instead of having a shadow in the wrong place, they just
have no shadow at all.