Creatures can run on two legs. Separated 'sprint' and 'run'.
The old quadrapedal run is now 'sprint', and the new bipedal run is 'run'.
creature-visuals.gd now stores these states in a 'movement_mode' enum.

Added 'elevation' flag so that shadows are drawn at an appropriate
distance when the creature is moving. The 'sitting on stools' logic also
uses this flag.

Removed unused creature-sprite.gd script.