New settings: Ghost piece, master volume. Back button fixes.
Ghost piece can be toggled with a setting. It defaults to 'on', and if
it's on a shadow appears at the piece's hard drop destination.

Master volume can be toggled, which affects the volume of
sfx/music/voices. I've changed the default volume from 0.5 to 0.7 across
the board, to compensate for the fact that all of the different audio
buses are affected by the master volume.

Fixed broken back button shortcuts. These were broken during the recent
touchscreen UI changes.

Made piece shadows slightly darker so that ghost piece would be more
visible.

Closes #485, #467.