rank-mode-demo.gd prints additional statistics.
I spent a part of this morning trying to rationalize the huge jump in
difficulty from 7d to 8d. The jump in difficulty is caused by the piece
speed change to 'FA' which allows for much higher play speeds, which
you're graded against.

This isn't necessarily a problem, I think it's OK for there to be a big
wall between 7d and 8d. We could even color-code it or make it clear 'this
is the wall' when you're in the game, or the ranks could have names like
'1-7 dan, 1-3 master, demon' or whatever we want