Added tutorial button. Fixed tutorial star seed bug.
Added a tutorial button to the main menu for testing tutorial
functionality. This is a placeholder, and will eventually behave
differently and possibly not be located on the main menu.

The tutorial's star seeds were lingering from one part of the tutorial to
the next. I've removed the logic for spawning the star seeds at the start of a
level because it was the easiest way to prevent these bugs.