U-piece has new rotation rules to facilitate climbing.
The U-piece used to pivot around its long straight part which is intuitive
because it's close to its center of mass. However, this makes it very
rigid and inflexible. I've changed it to pivot around its 'hole' which while
allows for more creative rotation options. Most importantly it allows the
U piece to climb up stacks of pieces, but also allows for rapid snack
boxes in the center column.

These changes are consistent with the P and Q pieces, which also initially
used a 'center of mass' rotation rule but were changed to allow for creative
rotation options.