Added new 'smirk' mouth. Repositioned emote arms.
New mouth looks good with a nose, so it's more likely to be paired with a
nose for random creatures. The new mouth smiles/frowns based on mood.

Repositioned arms when emoting; the arms used to be positioned to be
touching the creature's beak, which looked strange for non-beaked
creatures as the arm was floating in mid-air. It's now touching the
creature's chin.

Bort now has a nose and a different mouth.