Fixed emote bugs related to arms, eyes, orientation
Creatures who went from running to emoting had an extra arm, which was the
result of the 'idle' movement animation playing. The idle animation is now
stopped when emoting.

If Spira talked to a running creature, the creature wouldn't stop and
orient themselves correctly because the 'orient_toward' method said not
to. The orient_toward method now always orients a creature, and there's
some special logic in overworld-ui.gd which says that Spira is exempt from
orienting.

Fixed bug where some emotes would result in creatures having two eyes;
this was introduced in d6df127 when trying to fix some bugs with
north/south idle animations