Tutorial bugs: star seeds, instructor, music volume, messages
Fixed bug where star seeds sometimes lingered during tutorial. This bug was
caused because the after_piece_written and lines_deleted signals were emitted in
the wrong order. This meant when changing tutorial scenes, the scene would
change, and then the new star seeds would get moved around based on the cleared
lines.

Star seeds for incomplete boxes (such as in the third part of the tutorial) are
no longer placed adjacent if possible.

Instructor is now summoned when restarting a tutorial. Before, if you restarted
a tutorial after customers showed up, the random customer would lead the tutorial.

Increased tutorial music volume. The tutorial music is about 4 db quieter
than other songs, so I've turned up its in-game volume to compensate.
It was kind of jarring when the 'real music' kicked in before, because it
was a lot louder.

Fixed bug where message would sometimes be blanked out when restarting a
tutorial. This was because the pop_out_timer wasn't reset, so a message
might be marked to pop out, and then an important message appears
but disappears before it can be read.