Can change player's appearance, name.
The appearance and name is governed by a save file; there's no way to edit the
save file in-game but you can modify it in a text editor. The main character can
have different faces/colors/weight.

Substitutes player's name in conversations.

Changed 'Spira' to 'player' globally, now that the player's name is a
variable.

Updated chat theme colors to match character colors.