Added main character to restaurant scene.
CreatureView is now RestaurantView. CreatureView had a parent/child relationship
with its RestaurantScene and did not use a Viewport or Camera. It's been
rewritten to RestaurantView which utilizes viewports and cameras so that we can
have multiple viewports into the same scene.

Despite utilizing viewports, the quality is still the same as it was
before even at high resolutions. This was achieved by having
Viewports rendering things at 200% size, which are then scaled down in
ViewportContainers. We should be able to use the same technique to improve
the quality of shadows and puzzle fruits later.

Chef emotes when player does well, tops out, or finishes a level.

Moods transition into themselves better; before, when transitioning from
'smile0' to 'smile0' the chef would briefly open their eyes.