Separated creature chin to avoid layer issue with cheek/mouth
The creature mouth sprite used to include the actual beak/orifice/teeth and
stuff, as well as the chin. This caused problems because when the beak is
open it needs to overlap the cheek. But the chin needs to go beneath the
cheek.

This commit splits the chin into its own resource, and it's animated
independently. All mouths share the same chin; I tried to create a generic
set of chins which could work for any mouth.

Also moved some cheeks frames from z2 to z1 to appear behind the mouth.