Fixed thin shader lines at boundaries, overlaps.
This was noticable at the top of the neck as described in #492. But, it
also became noticable around the eyes and face after facial colors were
introduced recently.

This bug was caused because red/green/blue are mixed one after the other
-- but the old math would mix in red 33%, green 33% and blue 33% for
example, resulting in a color which still had some black in it. The new
logic will mix in red 100%, green 50%, and blue 33% for an opaque color
which is evenly distributed between all three colors.