Creature now only has one ear/mouth AnimationPlayer at once.
Instead of having three AnimationPlayers and turning them on/off, it now only
has one. This solution scales better, and we'll eventually need to do something
similar for bodies so it's helpful to have the framework in place.

CreatureLoader hosts the preloaded list of AnimationPlayers. These cannot go in
CreatureVisuals or it causes a circular reference. CreatureLoader is now an
autoloaded node instead of a static class.