Added framework for multiple body types. Split up CreatureVisuals.
There are two body types; body "1" is the normal body all creatures have,
body "2" is a placeholder square body  which demonstrates a proof of
concept. (This looks surprisingly passable and maybe we could do some lazy
variations like this for a robot creature or something like that.)

To facilitate swapping CreatureCurve and AnimationPlayer instances, they've been
split into their own scenes. I also split out EmotePlayer, because after
extracting the other stuff EmotePlayer's animations made up for 80% of the
remaining content in CreatureVisuals.tscn. It made sense to be its own thing.

Fixed floating point bugs in CreatureCurve.save_curve. It was relying on
'>' and '=' for looping/business logic which is unreliable for floats, and
caused curves to be inserted/replaced incorrectly.