Refactored CreatureVisuals methods, signals, repetitive code.
Split CreatureVisuals._update_creature_properties into _unassign and
_assign methods. These are slightly more descriptive and help separate and
document the two distinct pieces of functionality the method was doing.

Extracted _add_visuals_node, _remove_visuals_node utility methods to cut
down on repetitive code.

The creature_arrived signal is now named 'dna_loaded' which is more
descriptive. This was originally named 'creature_arrived' because it coincided
with arriving at the restaurant, but now that is not always the case as it
applies to creatures being loaded in the overworld or elsewhere.

I removed the before_creature_arrived signal. It was only used for stopping
animations which is now unnecessary as the animation players get removed.