ChatNametagPanel, TEXTURE_SCALE refactorings
Pushed NametagPanel.show_label() into new ChatNametagPanel class. This
show_label() function positioned a label around a chat frame, but the nametag
label is also used in the creature editor which has nothing to do with chat
frames (and will eventually be used elsewhere as well.) It's better for the
show_label function to be on a chat-specific class

Extracted Creature.TEXTURE_SCALE for shadows. Shadows need to be proportional to
the texture scale. I tried having the shadows reference the TextureRect.scale
directly but this didn't quite work because of initialization order (TextureRect
is nil when the shadow is initialized)