Added bird beak, cat ears. Added 'illegal allele combo' logic.
Creature editor now respects 'illegal combos'

When tweaking a row, if you click the dropdown, it will remove any choices
which result in illegal combinations. If you click the DNA symbol, it will
generate creatures which resolve the illegal combinations.

When mutating a creature, the mutate algorithm detects alleles which
cannot be mutated and picks different alleles. It also detects values
which would not be suitable and picks different values.