Fixed issue where playfield lights were not always initialized properly.
There was some logic which checked if the tileset was initialized, and if
it was already initialized, then the _color_tile_indexes map was left
empty. This logic was flawed because tilesets are sometimes cached if not
unique to a scene, but the _color_tile_indexes map needs to be initialized
every time.

These two objects are now initialized independently.

Closes #622.