Added squish tutorial.
This also includes logic for displaying diagrams, and having
non-interactive levels. I'd also like for this tutorial to steer some pieces
into place, but for now parts 5 and 6 just have the pieces already in place.

Lines and money are earned during tutorial now. We want to keep track of
statistics like how many lines the player's cleared or what their combo
is, so this is the most practical way to track it.

Removed finish conditions for basic tutorial. We use the idea of finish
conditions for piece-limited sections of tutorials, but we don't want to
risk people actually reaching them and messing up the tutorial.