Fixed PackedSprite editor errors, improved singleton logic
PackedSprite was throwing errors because it depended on ResourceCache
which was not a tool. ResourceCache shouldn't be a tool because caching
resources in the editor has the potential to cause headaches, as resources
change frequently during development. PackedSprite now has the capability
of loading non-cached resources in the editor.

Improved singleton logic to allow for further singletons. Instead of
specifically saving and loading a wallpaper singleton, ResourceCache now saves
and loads all nodes in the 'singletons' node group.