Added combo indicators which show +3, +4, +5...
This includes a font, a little backdrop icon and some particle effects.

Added fallback so that PackedSprite can still load resources if the
ResourceCache is not populated.

Changed Z-index of puzzle components such as food items and the active
piece. This fixes a bug where the active piece would appear behind star
seeds if the player topped out with the piece overlapping a box.